Add updated changes from 2016 Concept-C#
Also make modifications to allow Concept-C# to work with modern
Roslyn and .NET Framework.

I've merged this work into one commit to make keeping
this in sync with Roslyn upstream easier.  The full commit history
including attribution is available in the '2016' branch.

This commit includes a lot of work by @crusso, not just by me.