Concept-C#'17: Add manual overlap attributes
The concept inferrer will no longer tie-break unless directed to
by the existence of these GHC-style attributes.

This is in preparation for overhauling the tie-breaker to be a
lot more rigorous.

I think this is a little too heavy-handed at the moment.  There
should be a heuristic for when the tie is between two instances
related by inheritance, maybe.