Concept-C#'17: Concept extension methods
First implementation of 'concept extension methods', or CEMs:
the ability to mark up a concept method with extension method
syntax and have it considered like an extension method on the
first parameter.

CEMs are implemented under the hood with a new attribute,
several hooks into the extension method binding system,
and lots of hot-glue and duct-tape.