Concept-C#'17: Recursive MTI and CWI refactoring
A large battery of changes to the concept witness inferrer, both
refactoring it and also allowing it to call upon information about
its parent method when used in method position.

Currently we use this to invoke a recursive round of ordinary
Method Type Inference every now and then to fix dependencies
between eg. associated types in input position and parameters of
concepts in output position.

There are still some unsolved consistency issues: for example,
the inferrer can pick up one instance when a parameter is fixed
and another instance when the parameter is left unfixed.