Concept-C#'17: Quickfix cyclic dependency
Having a concept with missing TPs in a constraint would sometimes cause
an infinite loop of repeatedly getting the concept as an interface, part
inferring, getting the concept as an interface, and so on.

This quickfix disables part-inference in situations where this could
happen.  It might be incomplete, but should be sound.